--- dice.js
@@ -1,5 +1,0 @@
-function dice(number) {
-    return Math.floor(Math.random() * number + 1)
-}
-
-module.exports = { dice }
