+++ src/api/controllers/gamesController.ts
@@ -1,0 +1,5 @@
+function dice(steps : number) {
+    return Math.floor(Math.random() * steps + 1)
+}
+
+module.exports = { dice }
