+++ src/api/games/index.ts
@@ -1,8 +1,7 @@
-const express = require('express')
 const router = express.Router()
 
 router.get('/dice/:n',(req,res) => {
-    const number = dice.dice(req.params.n)
+    const number = dice(req.params.n)
     const result = { result : number }
     res.json(result)
 })
