+++ src/api/users/index.ts
@@ -1,8 +1,4 @@
-const express = require('express')
-const dice = require('../../../dice')
-const router = express.Router()
 
-//CREATE REQUEST UPDATE DELETE 
 const users = [{id: 0,name:'Pepe'},{id: 1,name:'Jose'}]
 
 
@@ -13,8 +9,6 @@ router.get('/', (req, res ) => {
 
 router.get('/:id', (req,res ) => {
     const userId = req.params.id
-    console.log(userId)
-    console.log(users[0].id)
     const user = users.find(user => user.id == userId )
     console.log(user)
     res.json(user)
